exclude password from fetch
