+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import { Suspense } from 'react';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import '@/styles/globals.css';
-import CustomHead from '@/components/CustomHead';
+import CustomHead from '@/components/CustomHead/CustomHead';
 
 export default function RootLayout({
   children,
