+++ src/components/CustomHead/index.ts
@@ -1,0 +1,1 @@
+export { default as CustomHead } from './CustomHead';
